requirements.txt
pyrogram
tgcrypto
pymongo
motor
aiohttp
google-generativeai

